// src/api/gen/types/ComponentDescriptor.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { ComponentMark } from './ComponentMark'
import type { ComponentRegion } from './ComponentRegion'

export type ComponentDescriptor = {
    component?: string;
    groupings?: string[];
    regions?: ComponentRegion[];
    vocabulary?: ComponentMark[];
};