// src/api/gen/types/MarkingOrigin.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export const markingOriginSourceEnum = {
    SESSION: "SESSION",
    HISTORY: "HISTORY"
} as const;

export type MarkingOriginSourceEnumKey = (typeof markingOriginSourceEnum)[keyof typeof markingOriginSourceEnum];

export type MarkingOrigin = {
    /**
     * @description
     * Format: `uuid`
     * @type string | undefined
    */
    encounterId?: string;
    practitionerName?: string;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    recordedAt?: string;
    /**
     * @description
     * Format: `date-time`
     * @type string | undefined
    */
    since?: string;
    source?: MarkingOriginSourceEnumKey;
};